Add files via upload
Hi my name is David and I am search firmware for ender 3 with SKR 1.4 turbo but no result. I am watching to much videos and google something to make this. This code have TMC2208, Bltouch working with pins servo and probe. I thing is a good for me maybe for you.